Fix createCampaign ABI mismatch with deployed contract
- Fix ABI parameter names: _description -> _desc, _targetAmount -> _fundingGoal
- Fix parameter types: uint32 -> uint256 for _fundingGoal
- Update CreateCampaign.tsx to use BigInt for uint256 parameters
- Match ABI with successful transaction parameters

This fixes the createCampaign function to match the actual deployed contract
that was used in the successful transaction 0xaa3a090ab8b462d6076472431e4ccbf3f80623e2d8632ed8231d5a2ae1b1aede

+++ components/CreateCampaign.tsx
@@ -30,9 +30,9 @@ export default function CreateCampaign() {
     setIsLoading(true);
 
     try {
-      // Use target amount as uint32 (in USD cents for precision)
-      const targetAmountInCents = parseInt(formData.targetAmount) * 100; // Convert USD to cents
-      const durationInSeconds = parseInt(formData.duration) * 24 * 60 * 60; // Convert days to seconds
+      // Use target amount as uint256 (in USD cents for precision)
+      const targetAmountInCents = BigInt(parseInt(formData.targetAmount) * 100); // Convert USD to cents
+      const durationInSeconds = BigInt(parseInt(formData.duration) * 24 * 60 * 60); // Convert days to seconds
 
       await write({
         address: CHARITY_NEXUS_ADDRESS,
@@ -41,8 +41,8 @@ export default function CreateCampaign() {
         args: [
           formData.name,
           formData.description,
-          targetAmountInCents, // uint32 type for target amount
-          BigInt(durationInSeconds),
+          targetAmountInCents, // uint256 type for target amount
+          durationInSeconds,
         ],
       });
